pyproject.toml
[project]
name = "word-para-excel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.9"
dependencies = [
    "document>=1.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "python-docx>=1.2.0",
]
